+++ src/components/ChannelList/hooks/usePaginatedChannels.ts
@@ -9,6 +9,7 @@ import { useChatContext } from '../../../context/ChatContext';
 
 import type { DefaultStreamChatGenerics } from '../../../types/types';
 import type { ChannelsQueryState } from '../../Chat/hooks/useChannelsQueryState';
+import { DEFAULT_INITIAL_CHANNEL_PAGE_SIZE } from '../../../constants/limits';
 
 const RECOVER_LOADED_CHANNELS_THROTTLE_INTERVAL_IN_MS = 5000;
 const MIN_RECOVER_LOADED_CHANNELS_THROTTLE_INTERVAL_IN_MS = 2000;
@@ -79,6 +80,7 @@ export const usePaginatedChannels = <
 
         const newOptions = {
           limit: options?.limit ?? MAX_QUERY_CHANNELS_LIMIT,
+          message_limit: options?.message_limit ?? DEFAULT_INITIAL_CHANNEL_PAGE_SIZE,
           offset,
           ...options,
         };