+++ src/components/MessageList/VirtualizedMessageList.tsx
@@ -53,6 +53,7 @@ import { ComponentContextValue, useComponentContext } from '../../context/Compon
 
 import type { Channel, ChannelState as StreamChannelState, UserResponse } from 'stream-chat';
 import type { DefaultStreamChatGenerics, UnknownType } from '../../types/types';
+import { DEFAULT_NEXT_CHANNEL_PAGE_SIZE } from '../../constants/limits';
 
 type VirtualizedMessageListPropsForContext =
   | 'additionalMessageInputProps'
@@ -65,6 +66,7 @@ type VirtualizedMessageListPropsForContext =
   | 'messageActions'
   | 'shouldGroupByUser'
   | 'sortReactions'
+  | 'sortReactionDetails'
   | 'threadList';
 
 /**
@@ -184,7 +186,7 @@ const VirtualizedMessageListWithContext = <
     loadMoreNewer,
     Message: MessageUIComponentFromProps,
     messageActions,
-    messageLimit = 100,
+    messageLimit = DEFAULT_NEXT_CHANNEL_PAGE_SIZE,
     messages,
     notifications,
     // TODO: refactor to scrollSeekPlaceHolderConfiguration and components.ScrollSeekPlaceholder, like the Virtuoso Component
@@ -196,6 +198,7 @@ const VirtualizedMessageListWithContext = <
     separateGiphyPreview = false,
     shouldGroupByUser = false,
     showUnreadNotificationAlways,
+    sortReactionDetails,
     sortReactions,
     stickToBottomScrollBehavior = 'smooth',
     suppressAutoscroll,
@@ -385,7 +388,9 @@ const VirtualizedMessageListWithContext = <
     }
   };
   const atTopStateChange = (isAtTop: boolean) => {
-    if (isAtTop) loadMore?.(messageLimit);
+    if (isAtTop) {
+      loadMore?.(messageLimit);
+    }
   };
 
   useEffect(() => {
@@ -447,6 +452,7 @@ const VirtualizedMessageListWithContext = <
               ownMessagesReadByOthers,
               processedMessages,
               shouldGroupByUser,
+              sortReactionDetails,
               sortReactions,
               threadList,
               unreadMessageCount: channelUnreadUiState?.unread_messages,
@@ -493,7 +499,8 @@ type PropsDrilledToMessage =
   | 'additionalMessageInputProps'
   | 'customMessageActions'
   | 'messageActions'
-  | 'sortReactions';
+  | 'sortReactions'
+  | 'sortReactionDetails';
 
 export type VirtualizedMessageListProps<
   StreamChatGenerics extends DefaultStreamChatGenerics = DefaultStreamChatGenerics