+++ src/components/MessageList/VirtualizedMessageListComponents.tsx
@@ -138,6 +138,7 @@ export const messageRenderer = <
     ownMessagesReadByOthers,
     processedMessages: messageList,
     shouldGroupByUser,
+    sortReactionDetails,
     sortReactions,
     unreadMessageCount = 0,
     UnreadMessagesSeparator,
@@ -212,6 +213,7 @@ export const messageRenderer = <
         Message={MessageUIComponent}
         messageActions={messageActions}
         readBy={ownMessagesReadByOthers[message.id] || []}
+        sortReactionDetails={sortReactionDetails}
         sortReactions={sortReactions}
       />
       {showUnreadSeparatorBelow && (